voice-demo/requirements.txt
Flask==3.0.0
openai-whisper==20231117
dashscope==1.14.0
numpy==1.24.0
sounddevice==0.4.6
soundfile==0.12.1
